requirements.txt
boto3==1.35.89
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
playwright==1.49.0
aiohttp==3.11.11

